fix defence fix

+++ Assets/02_Scripts/Utilities/DamageCalculator.cs
@@ -144,7 +144,7 @@ public static class DamageCalculator
 
         result = elementalDef * baseDamage;//속성 대미지 계산
         //버프 디버프에 의한 대미지 계산 추가
-        result = result * (1 - target.Defence / (50 + target.Defence));//방어력 계산
+        result = result * (1 - target.Defence / (50f + (float)target.Defence));//방어력 계산
         Debug.Log($"최종 대미지 : {result}");
 
         return result;
